requirements.txt
uvicorn 
fastapi
httpx
jinja2
python-multipart
ollama
loguru
transformers
ddgs
watchfiles
reportlab
pdfkit
wkhtmltopdf
sqlalchemy
load_dotenv
asyncpg
supervision
opencv-python
werkzeug
python-jose[cryptography]
passlib
pydantic-settings
bcrypt
pydantic[email]
valkey[libvalkey]
pytest
pytz
redis
